requirements.txt
streamlit>=1.24.0
selenium>=4.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
openpyxl>=3.1.0
unidecode>=1.3.6
phonenumbers>=8.13.0
loguru>=0.7.0
numbers-parser>=3.0.0